Renamed Emu/GS -> Emu/RSX
