+++ rpcs3/Emu/RSX/GL/OpenGL.h
