+++ rpcs3/Emu/RSX/GSManager.h
