+++ rpcs3/Emu/RSX/RSXTexture.h
