+++ rpcs3/Emu/RSX/RSXVertexProgram.h
