+++ rpcs3/Emu/RSX/sysutil_video.h
