+++ rpcs3/emucore.vcxproj
@@ -74,22 +74,22 @@
     <ClCompile Include="Emu\FS\vfsLocalFile.cpp" />
     <ClCompile Include="Emu\FS\vfsStream.cpp" />
     <ClCompile Include="Emu\FS\vfsStreamMemory.cpp" />
-    <ClCompile Include="Emu\GS\GL\GLBuffers.cpp" />
-    <ClCompile Include="Emu\GS\GL\GLFragmentProgram.cpp" />
-    <ClCompile Include="Emu\GS\GL\GLGSRender.cpp" />
-    <ClCompile Include="Emu\GS\GL\GLProgram.cpp" />
-    <ClCompile Include="Emu\GS\GL\GLProgramBuffer.cpp" />
-    <ClCompile Include="Emu\GS\GL\GLVertexProgram.cpp" />
-    <ClCompile Include="Emu\GS\GL\OpenGL.cpp" />
-    <ClCompile Include="Emu\GS\GSManager.cpp" />
-    <ClCompile Include="Emu\GS\GSRender.cpp" />
-    <ClCompile Include="Emu\GS\RSXTexture.cpp" />
-    <ClCompile Include="Emu\GS\RSXThread.cpp" />
     <ClCompile Include="Emu\HDD\HDD.cpp" />
     <ClCompile Include="Emu\Io\Keyboard.cpp" />
     <ClCompile Include="Emu\Io\Mouse.cpp" />
     <ClCompile Include="Emu\Io\Pad.cpp" />
     <ClCompile Include="Emu\Memory\Memory.cpp" />
+    <ClCompile Include="Emu\RSX\GL\GLBuffers.cpp" />
+    <ClCompile Include="Emu\RSX\GL\GLFragmentProgram.cpp" />
+    <ClCompile Include="Emu\RSX\GL\GLGSRender.cpp" />
+    <ClCompile Include="Emu\RSX\GL\GLProgram.cpp" />
+    <ClCompile Include="Emu\RSX\GL\GLProgramBuffer.cpp" />
+    <ClCompile Include="Emu\RSX\GL\GLVertexProgram.cpp" />
+    <ClCompile Include="Emu\RSX\GL\OpenGL.cpp" />
+    <ClCompile Include="Emu\RSX\GSManager.cpp" />
+    <ClCompile Include="Emu\RSX\GSRender.cpp" />
+    <ClCompile Include="Emu\RSX\RSXTexture.cpp" />
+    <ClCompile Include="Emu\RSX\RSXThread.cpp" />
     <ClCompile Include="Emu\SysCalls\Callback.cpp" />
     <ClCompile Include="Emu\SysCalls\FuncList.cpp" />
     <ClCompile Include="Emu\SysCalls\lv2\lv2Fs.cpp" />
@@ -284,22 +284,6 @@
     <ClInclude Include="Emu\FS\vfsStream.h" />
     <ClInclude Include="Emu\FS\vfsStreamMemory.h" />
     <ClInclude Include="Emu\GameInfo.h" />
-    <ClInclude Include="Emu\GS\GCM.h" />
-    <ClInclude Include="Emu\GS\GL\GLBuffers.h" />
-    <ClInclude Include="Emu\GS\GL\GLFragmentProgram.h" />
-    <ClInclude Include="Emu\GS\GL\GLGSRender.h" />
-    <ClInclude Include="Emu\GS\GL\GLProgram.h" />
-    <ClInclude Include="Emu\GS\GL\GLProgramBuffer.h" />
-    <ClInclude Include="Emu\GS\GL\GLShaderParam.h" />
-    <ClInclude Include="Emu\GS\GL\GLVertexProgram.h" />
-    <ClInclude Include="Emu\GS\GL\OpenGL.h" />
-    <ClInclude Include="Emu\GS\GSManager.h" />
-    <ClInclude Include="Emu\GS\GSRender.h" />
-    <ClInclude Include="Emu\GS\Null\NullGSRender.h" />
-    <ClInclude Include="Emu\GS\RSXFragmentProgram.h" />
-    <ClInclude Include="Emu\GS\RSXTexture.h" />
-    <ClInclude Include="Emu\GS\RSXThread.h" />
-    <ClInclude Include="Emu\GS\RSXVertexProgram.h" />
     <ClInclude Include="Emu\HDD\HDD.h" />
     <ClInclude Include="Emu\Io\Keyboard.h" />
     <ClInclude Include="Emu\Io\KeyboardHandler.h" />
@@ -314,6 +298,24 @@
     <ClInclude Include="Emu\Memory\DynamicMemoryBlockBase.h" />
     <ClInclude Include="Emu\Memory\Memory.h" />
     <ClInclude Include="Emu\Memory\MemoryBlock.h" />
+    <ClInclude Include="Emu\RSX\GCM.h" />
+    <ClInclude Include="Emu\RSX\GL\GLBuffers.h" />
+    <ClInclude Include="Emu\RSX\GL\GLFragmentProgram.h" />
+    <ClInclude Include="Emu\RSX\GL\GLGSRender.h" />
+    <ClInclude Include="Emu\RSX\GL\GLProcTable.h" />
+    <ClInclude Include="Emu\RSX\GL\GLProgram.h" />
+    <ClInclude Include="Emu\RSX\GL\GLProgramBuffer.h" />
+    <ClInclude Include="Emu\RSX\GL\GLShaderParam.h" />
+    <ClInclude Include="Emu\RSX\GL\GLVertexProgram.h" />
+    <ClInclude Include="Emu\RSX\GL\OpenGL.h" />
+    <ClInclude Include="Emu\RSX\GSManager.h" />
+    <ClInclude Include="Emu\RSX\GSRender.h" />
+    <ClInclude Include="Emu\RSX\Null\NullGSRender.h" />
+    <ClInclude Include="Emu\RSX\RSXFragmentProgram.h" />
+    <ClInclude Include="Emu\RSX\RSXTexture.h" />
+    <ClInclude Include="Emu\RSX\RSXThread.h" />
+    <ClInclude Include="Emu\RSX\RSXVertexProgram.h" />
+    <ClInclude Include="Emu\RSX\sysutil_video.h" />
     <ClInclude Include="Emu\SysCalls\Callback.h" />
     <ClInclude Include="Emu\SysCalls\ErrorCodes.h" />
     <ClInclude Include="Emu\SysCalls\LogBase.h" />
